pyproject.toml
[tool.poetry]
name = "notion-automation"
version = "0.1.0"
description = ""
authors = ["Patrick Black <info@patrickblackjr.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
